setting up TailMetricsServiceImpl
